+++ src/components/common/Tabs/index.js
@@ -17,10 +17,11 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const Tabs = memo(({ tabs = [] }) => {
+const Tabs = memo(({ tabs = [], ...props }) => {
   const { t } = useTranslation();
   const classes = useStyles();
   const [tabNumber, setTabNumber] = useState(0);
+  const Content = tabs[tabNumber].component;
 
   return (
     <Box boxShadow={3}>
@@ -33,7 +34,7 @@ const Tabs = memo(({ tabs = [] }) => {
           {tabs.map(({ label }) => (<AntTab key={label} label={t(label)} />))}
         </AntTabs>
         <div className={classes.content}>
-          {tabs[tabNumber].component}
+          <Content {...props} />
         </div>
       </div>
     </Box>
