+++ src/pages/Pet.js
@@ -1,8 +1,8 @@
-import PetContent from 'modules/Pet';
+import { EditPet } from 'modules/Pet';
 import { PetProvider } from 'modules/Pet/context';
 
 export default () => (
   <PetProvider>
-    <PetContent />
+    <EditPet />
   </PetProvider>
 );
